Revert "update my version to Josh's new version"
