+++ KeyCrush.py
@@ -1,10 +1,9 @@
 from random import choice, randint
-from BELLA_GAME import Bella_Game
 
-class KeyCrush(Bella_Game):
+
+class KeyCrush:
 
     def __init__(self, gametools, display_data, starting_level=0):
-        super().__init__(gametools, display_data, starting_level)
         """
 
         self.alphabet: list, letters to be tested, in order.
@@ -13,11 +12,11 @@ class KeyCrush(Bella_Game):
         
         self.letters_correct: list, int between 0 and self.max_correct.
 
-        self.game_state: string; introduction, test_leter, test_word,
-                                 or announcement.
-
         self.max_correct: int, maximum number of times letter must be
                           answered corectly to advance a level.
+
+        self.game_state: string; introduction, test_leter, test_word,
+                                 or announcement.
 
         self.level: int, the current player's level.
 
@@ -35,9 +34,6 @@ class KeyCrush(Bella_Game):
 
         self.input_letter: string, the letter most recently typed.
 
-        self.input_control = string, holds input from keyboard (used
-                             here to change display and receive space bar)
-
         self.total_attempted_letters_answered: int, total number of attempted
                                      answers.
                                      
@@ -52,11 +48,11 @@ class KeyCrush(Bella_Game):
         self.letter_attempts_before_word: int, tallies answered letters;
                                           helps switch to word if above some #.
 
-        self.letter_attempts_before_hint = int, number of wrong answers for current letter
-                                           (after 2, a hint is given on subsequent ones)
-
         self.letter_attempts_threshold: int, number of attempts before possible
                                         to switch to word.
+                                
+        self.letter_attempts_before_hint: int, tallies incorrect responses;
+                                          gives hint if above some #.
                                           
         self.total_words_answered: int, total number of words answered.
         
@@ -71,27 +67,66 @@ class KeyCrush(Bella_Game):
                           
         self.word_prompt: string, the word to be answered.
 
-        self.word_has_been_said = boolean, whether current word soundfile has been
-                                  played by game yet.
-
         self.intro_done: boolean, has the introduction sound been played?
-
-        self.points_to_be_awareded: int, base number of points, adjusted depending on
-                                    whether earned in a streak or after a wrong letter, etc.
-
-        self.total_points: int, tally of all points so far in game.
-
-        self.using_card: boolean, whether game is using an inserted card or the game's own
-                         alphabet.
-        
-        self.card_str: string, the set of letters read from an inserted card.  If a card is
-                       being used, the letters to be tested are chosen from tis, rather than
-                       the game's self.alphabet string.
                 
         """
 
+#---META-GAME STUFF---
+
+        self.pygame = gametools['pygame']
+        self.sounds = gametools['sounds']
+        self.np = gametools['numpy']
+        self.gameDisplay = gametools['display']
+        self.braille_keyboard = gametools['keyboard']
+
+        self.sound_object = self.sounds.sounds()
+        self.game_name = 'KeyCrush'
+
+#---DISPLAY---
+        
+        self.SCREEN_WIDTH = display_data['screen_width']
+        self.SCREEN_HEIGHT = display_data['screen_height']
+
+        self.pygame.display.set_caption('Typing Tutor')
+   
+        self.font = self.pygame.font.SysFont(None, 80)
+        self.font_small = self.pygame.font.SysFont(None, 40)
+        self.font_large = self.pygame.font.SysFont(None, 500)
+        
+        self.white, self.black, self.yellow, self.blue = (255, 255, 255), (0, 0, 0), (255, 255, 0), (0, 0, 255)
+
+        self.current_display_state = display_data['current_display_state']
+
+        self.display_names = ['white_black', 'black_white', 'blue_yellow']
+
+        self.display_states = {'black_white':{'background':self.black, 'text':self.white},
+                               'white_black':{'background':self.white, 'text':self.black},
+                               'blue_yellow':{'background':self.blue, 'text':self.yellow}}
+
+
+#---SOUNDS---
+
+        standard_alphabet_dir= self.sounds.join('standardsounds', 'Alphabet')
+        standard_sfx_dir = self.sounds.join('standardsounds', 'Sfx')
+        standard_voice_dir = self.sounds.join('standardsounds', 'Voice')
+
+
+        self.standard_alphabet = self.sound_object.make_sound_dictionary(standard_alphabet_dir, self.pygame)
+        self.standard_sfx = self.sound_object.make_sound_dictionary(standard_sfx_dir, self.pygame)
+        self.standard_voice = self.sound_object.make_sound_dictionary(standard_voice_dir, self.pygame)
+        
+        self.game_sounds = self.sound_object.make_sound_dictionary(self.game_name + '_sounds', self.pygame)
+
+        
+        self.standard_alphabet[' '] = {'sound':self.pygame.mixer.Sound(self.sounds.join(standard_alphabet_dir, 'space.wav')),
+                                     'length':int(self.pygame.mixer.Sound(self.sounds.join(standard_alphabet_dir, 'space.wav')).get_length() * 1000)}
+
+       
 #---GAME VARIABLES---
 
+        self.correct_sfx = ["correct_one","correct_two","correct_three","correct_four","correct_five","correct_six"]
+        self.correct_voice = ["fantastic","keep_it_up","good_job","great_job","outstanding"]
+        
         self.alphabet = 'abcdefghijklmnopqrstuvwxyz'
 
         self.list_word_prompts = [["ace", "bad", "cab"],
@@ -143,28 +178,35 @@ class KeyCrush(Bella_Game):
 
     def iterate(self, input_dictionary):
         """ A single iteration of the game loop.
-            Runs either introduction, test_letter,
-            or test_word each time.
         """
 
-        self.gameDisplay.fill(self.display_states[self.display_names[self.current_display_state]]['background'])
-
-        self.input_letter = input_dictionary['letter']
         self.input_control = input_dictionary['standard']
         
         if (self.input_control == 'display'):
             self.change_display_state()
 
-        if input_dictionary['card_trigger']:
-            self.get_card_data()
+        self.input_letter = input_dictionary['letter']
+
+##        if input_dictionary['card_trigger']: #make this its own function
+##            self.using_card = True
+##            temp_string = input_dictionary['card_str']
+##            for character in temp_string:
+##                if (character in self.alphabet) & (character not in self.card_str):
+##                    self.card_str += character
+##            self.letters_correct = self.np.ones(len(self.card_str))
+##            self.switch_to_letter()
 
         if self.input_letter != None:
             self.play_sound('type', self.standard_sfx)
-
+        
+        self.gameDisplay.fill(self.display_states[self.display_names[self.current_display_state]]['background'])
+            
         if self.game_state == "introduction":
             self.introduction()
+
         elif self.game_state == "testing letter":
             self.test_letter()
+
         elif self.game_state == "testing word":
             self.test_word()
 
@@ -207,7 +249,14 @@ class KeyCrush(Bella_Game):
         """
 
         if self.word_has_been_said:
-            self.update_and_check_word()
+        
+            if self.input_letter != None:
+                self.word_string += self.input_letter
+            elif (len(self.word_string) >= len(self.word_prompt)) | (self.input_control == 'space'):
+                if self.word_string == self.word_prompt:
+                    self.word_is_correct()
+                else:
+                    self.word_is_wrong()
         else:
             self.play_sound(self.word_prompt, self.game_sounds)
             self.word_has_been_said = True
@@ -219,39 +268,9 @@ class KeyCrush(Bella_Game):
 
 #---OTHER FUNCTIONS---
 
-    def get_card_data(self):
-        """ Read data from card, check to make sure characters
-            are ones that KeyCrush would test--and also not already
-            in the string to be tested, and if so, include
-            them.
-        """
-        
-        self.using_card = True
-        temp_string = input_dictionary['card_str']
-        for character in temp_string:
-            if (character in self.alphabet) & (character not in self.card_str):
-                self.card_str += character
-        self.letters_correct = self.np.ones(len(self.card_str))
-        self.switch_to_letter()
-
-
-    def update_and_check_word(self):
-        """ Add the most recent typed letter to a string and check
-            whether the string as long as the word string or whether the
-            user has typed space.  If either is true, compare the words.
-        """
-        
-        if self.input_letter != None:
-            self.word_string += self.input_letter
-        elif (len(self.word_string) >= len(self.word_prompt)) | (self.input_control == 'space'):
-            if self.word_string == self.word_prompt:
-                self.word_is_correct()
-            else:
-                self.word_is_wrong() 
-
 
     def letter_is_correct(self):
-        """ Update variables for a correctly-answered letter,
+        """ Update variables for a correclty-answered letter,
             play response sound, and call for a new letter prompt.
         """
         
@@ -259,28 +278,27 @@ class KeyCrush(Bella_Game):
         self.letters_answered_correctly += 1
         self.letter_streak += 1
         self.total_points += self.letter_streak * self.points_to_be_awarded
-        self.letter_attempts_before_hint = 0
-        self.letter_attempts_before_word += 1
-
+        
         if self.letter_streak > self.career_streak:
             self.career_streak = self.letter_streak
             
-        self.update_letter_tracking(True)
+        self.letter_attempts_before_hint = 0
+        self.letter_attempts_before_word += 1
         
-        if self.letter_streak % 5 == 0:
-            self.play_streak_sound()
-        else:
-            self.play_sound('correct_coins', self.game_sounds)
-
+        self.update_letter_tracking(True)
         self.check_level()
+
+        if self.letter_streak % 5 == 0:
+            self.play_sound(choice(self.correct_sfx), self.standard_sfx, wait=True)
+            self.play_sound(choice(self.correct_voice),self.standard_voice, wait=True)
+        else:            
+            self.play_sound(choice(self.correct_sfx), self.standard_sfx)
 
         if (self.letter_attempts_before_word > self.letter_attempts_threshold) & (not self.using_card):
             self.gamble_switch_to_word()
         else:
             self.get_new_letter()
 
-    def play_streak_sound(self):
-        pass
 
     def letter_is_wrong(self):
         """ Update variables for an incorrectly-answered letter,
@@ -291,13 +309,14 @@ class KeyCrush(Bella_Game):
         self.letter_streak = 0
         self.letter_attempts_before_hint += 1
         self.points_to_be_awarded = int(self.points_to_be_awarded * 0.5)
-
-        if self.letter_attempts_before_hint > 2:
+        self.update_letter_tracking(False)
+        self.play_sound('wrong', self.standard_sfx)
+        
+        if self.letter_attempts_before_hint > 1:
             self.points_to_be_awarded = 0
             self.give_hint()
         
-        self.update_letter_tracking(False)
-        self.play_sound('wrong_buzz', self.game_sounds)
+        
 
         print('Wrong letter!')
 
@@ -310,8 +329,8 @@ class KeyCrush(Bella_Game):
         self.words_answered_correctly += 1
         self.word_streak += 1
         self.total_points += self.points_to_be_awarded
-        self.play_sound('hang_of_it', self.game_sounds, wait=True)
-
+        self.play_sound('correct_word', self.standard_sfx)
+        self.play_sound(choice(self.correct_voice),self.standard_voice, wait=True)
         self.switch_to_letter()
 
 
@@ -321,7 +340,7 @@ class KeyCrush(Bella_Game):
         """
         self.total_words_answered += 1
         self.word_streak = 0
-        self.play_sound('wrong_buzz', self.game_sounds)
+        self.play_sound('almost', self.standard_voice)
         self.switch_to_letter()
 
 
@@ -457,3 +476,107 @@ class KeyCrush(Bella_Game):
             several variables.
         """
         return self.total_points
+
+
+#---SOUND FUNCTIONS---
+    
+    def play_sound(self, sound, dictionary, wait=False):
+        dictionary[sound]['sound'].play()
+        if wait:
+            self.pygame.time.wait(dictionary[sound]['length'])
+
+
+#---DISPLAY FUNCTIONS---
+
+    def change_display_state(self):
+        self.current_display_state = (self.current_display_state + 1) % len(self.display_names)
+
+
+    def display_letter_prompt(self, letter=None):
+        """ Write the current letter prompt to the screen.
+        """
+        if letter == None:
+            letter = self.letter_prompt
+            
+        displaybox = self.pygame.draw.rect(self.gameDisplay,
+                                           self.display_states[self.display_names[self.current_display_state]]['background'],
+                                           ((self.SCREEN_WIDTH/2)-200, 108, 400, 50))
+
+        text = self.font_large.render(letter, True,
+                                      self.display_states[self.display_names[self.current_display_state]]['text'])
+
+        temp_width = text.get_rect().width
+
+        self.gameDisplay.blit(text, ((self.SCREEN_WIDTH / 2) - (temp_width/2), 100))
+
+
+    def display_word_prompt(self, word=None):
+        """ Write the current word prompt to the screen.
+        """
+
+        if word == None:
+            word = self.word_prompt
+        displaybox = self.pygame.draw.rect(self.gameDisplay, self.display_states[self.display_names[self.current_display_state]]['background'], ((self.SCREEN_WIDTH/2)-200, 108, 400, 50))
+        text = self.font.render(word, True, self.display_states[self.display_names[self.current_display_state]]['text'])
+        temp_width = text.get_rect().width
+        self.gameDisplay.blit(text, ((self.SCREEN_WIDTH / 2) - (temp_width/2), 100))
+
+
+    def display_status_box(self):
+        """ Write the current word prompt to the screen.
+        """
+        
+        text = self.font_small.render("Level: " + str(self.level), True,
+                                      self.display_states[self.display_names[self.current_display_state]]['text'],
+                                      self.display_states[self.display_names[self.current_display_state]]['background'])
+        
+        text2 = self.font_small.render("Points: " + str(self.score), True,
+                                       self.display_states[self.display_names[self.current_display_state]]['text'],
+                                       self.display_states[self.display_names[self.current_display_state]]['background'])
+        
+        temp_width = text.get_rect().width
+        self.gameDisplay.blit(text, ((self.SCREEN_WIDTH / 10) - (temp_width/2), 10))
+#        self.gameDisplay.blit(text2, ((self.SCREEN_WIDTH / 10) - (temp_width/2), 45))
+
+
+    def draw_buttons(self, keys='000000'):
+        """ Draw all six buttons to the screen.
+            Color depends on input code.
+        """
+        
+        xpos = None
+        ypos = 500
+        button_width = 60
+        button_height = 80
+        offset = 7
+
+        x_divisor = self.SCREEN_WIDTH / 6
+        x_scalar = 0.8
+        x_buffer = (1.0 - x_scalar) * self.SCREEN_WIDTH * 0.5
+        x_middle = self.SCREEN_WIDTH * 0.07
+
+        key_order = [3, 4, 5, 2, 1, 0]
+        
+        for i in range(len(keys)):
+            if keys[key_order[i]] == '1':
+                color = self.display_states[self.display_names[self.current_display_state]]['text']
+            elif keys[key_order[i]] == '0':
+                color = self.display_states[self.display_names[self.current_display_state]]['background']
+            
+            if i > 2:
+                xpos = (i* x_divisor * x_scalar) + x_buffer + x_middle
+            else:
+                xpos = (i* x_divisor * x_scalar) + x_buffer
+
+            position = (xpos, ypos, button_width, button_height)
+            position_small = (xpos + offset, ypos + offset, button_width - (2 * offset), button_height - (2 * offset))
+
+            self.draw_single_button(self.display_states[self.display_names[self.current_display_state]]['text'], position)
+            self.draw_single_button(color, position_small)
+
+    
+    def draw_single_button(self, color, position):
+        """ Draw a single button to the screen.
+        """
+        
+        self.pygame.draw.ellipse(self.gameDisplay, color, position)
